Summer 2025 Release
The summer 2025 patch updates a bunch of the old dates to modern ones, re-enables the projects tab, and most importantly implements my idea of a modular carousel announcement editor.

A new folder called context has been added to the source and a new page in that folder called announcementcontext. In the admin folder a new folder has been added called announcement admin with a file with the same name. Modifications to Admin, Navi, and App JS have all been changed for the new function. Further implementation of new modular functions and or patches/updates will be coming soon.
-JV